refactor!: remove deprecated GeneratedVaadinNotification
