+++ vaadin-notification-flow-parent/vaadin-notification-flow/pom.xml
@@ -21,6 +21,11 @@
         </dependency>
         <dependency>
             <groupId>com.vaadin</groupId>
+            <artifactId>vaadin-flow-components-base</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.vaadin</groupId>
             <artifactId>vaadin-renderer-flow</artifactId>
             <version>${project.version}</version>
         </dependency>
