+++ vaadin-notification-flow-parent/vaadin-notification-flow/src/main/java/com/vaadin/flow/component/notification/Notification.java
@@ -18,21 +18,26 @@ package com.vaadin.flow.component.notification;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.Stream.Builder;
 
 import com.vaadin.flow.component.AttachEvent;
 import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.ComponentUtil;
 import com.vaadin.flow.component.DetachEvent;
 import com.vaadin.flow.component.HasComponents;
+import com.vaadin.flow.component.HasSize;
 import com.vaadin.flow.component.HasStyle;
 import com.vaadin.flow.component.HasTheme;
+import com.vaadin.flow.component.Synchronize;
+import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.Text;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.dependency.JsModule;
+import com.vaadin.flow.component.dependency.NpmPackage;
+import com.vaadin.flow.component.shared.HasThemeVariant;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.dom.ElementFactory;
 import com.vaadin.flow.dom.Style;
@@ -48,10 +53,16 @@ import com.vaadin.flow.shared.Registration;
  *
  * @author Vaadin Ltd
  */
+@Tag("vaadin-notification")
+@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha6")
+@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
+@NpmPackage(value = "@vaadin/notification", version = "24.0.0-alpha6")
+@JsModule("@vaadin/notification/src/vaadin-notification.js")
+@JsModule("@vaadin/polymer-legacy-adapter/template-renderer.js")
 @JsModule("./flow-component-renderer.js")
 @JsModule("./notificationConnector.js")
-public class Notification extends GeneratedVaadinNotification<Notification>
-        implements HasComponents, HasTheme, HasStyle {
+public class Notification extends Component implements HasComponents, HasStyle,
+        HasThemeVariant<NotificationVariant> {
 
     private static final int DEFAULT_DURATION = 5000;
     private static final Position DEFAULT_POSITION = Position.BOTTOM_START;
@@ -180,7 +191,7 @@ public class Notification extends GeneratedVaadinNotification<Notification>
     public Notification(String text, int duration, Position position) {
         initBaseElementsAndListeners();
         setText(text);
-        setDuration((double) duration);
+        setDuration(duration);
         setPosition(position);
     }
 
@@ -282,7 +293,9 @@ public class Notification extends GeneratedVaadinNotification<Notification>
      *            not {@code null}
      */
     public void setPosition(Position position) {
-        setPosition(position.getClientName());
+        String positionName = position.getClientName();
+        getElement().setProperty("position",
+                positionName == null ? "" : positionName);
     }
 
     /**
@@ -302,7 +315,7 @@ public class Notification extends GeneratedVaadinNotification<Notification>
      * @return the {@link Position} property from the webcomponent
      */
     public Position getPosition() {
-        String position = getPositionString();
+        String position = getElement().getProperty("position");
         return Optional.ofNullable(position).map(Position::fromClientName)
                 .orElse(DEFAULT_POSITION);
     }
@@ -310,7 +323,6 @@ public class Notification extends GeneratedVaadinNotification<Notification>
     /**
      * Opens the notification.
      */
-    @Override
     public void open() {
         setOpened(true);
     }
@@ -321,7 +333,6 @@ public class Notification extends GeneratedVaadinNotification<Notification>
      * Note: This method also removes the notification component from the DOM
      * after closing it, unless you have added the component manually.
      */
-    @Override
     public void close() {
         setOpened(false);
     }
@@ -431,7 +442,6 @@ public class Notification extends GeneratedVaadinNotification<Notification>
      *            {@code true} to open the notification, {@code false} to close
      *            it
      */
-    @Override
     public void setOpened(boolean opened) {
         UI ui = UI.getCurrent();
         if (ui == null) {
@@ -464,52 +474,66 @@ public class Notification extends GeneratedVaadinNotification<Notification>
                         }
                     });
         }
-        super.setOpened(opened);
+        getElement().setProperty("opened", opened);
     }
 
     /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
      * True if the notification is currently displayed.
      * <p>
      * This property is synchronized automatically from client side when a
      * 'opened-changed' event happens.
-     * </p>
      *
      * @return the {@code opened} property from the webcomponent
      */
+    @Synchronize(property = "opened", value = "opened-changed")
     public boolean isOpened() {
-        return isOpenedBoolean();
-    }
-
-    @Override
-    public Registration addOpenedChangeListener(
-            ComponentEventListener<OpenedChangeEvent<Notification>> listener) {
-        return super.addOpenedChangeListener(listener);
+        return getElement().getProperty("opened", false);
     }
 
     /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
-     * <p>
+     * `opened-changed` event is sent when the notification opened state
+     * changes.
+     */
+    public static class OpenedChangeEvent extends ComponentEvent<Notification> {
+        private final boolean opened;
+
+        public OpenedChangeEvent(Notification source, boolean fromClient) {
+            super(source, fromClient);
+            this.opened = source.isOpened();
+        }
+
+        public boolean isOpened() {
+            return opened;
+        }
+    }
+
+    /**
+     * Adds a listener for {@code opened-changed} events fired by the
+     * webcomponent.
+     *
+     * @param listener
+     *            the listener
+     * @return a {@link Registration} for removing the event listener
+     */
+    public Registration addOpenedChangeListener(
+            ComponentEventListener<OpenedChangeEvent> listener) {
+        return getElement().addPropertyChangeListener("opened",
+                event -> listener.onComponentEvent(
+                        new OpenedChangeEvent(this, event.isUserOriginated())));
+    }
+
+    /**
      * The duration in milliseconds to show the notification. Set to {@code 0}
      * or a negative number to disable the notification auto-closing.
-     * </p>
      *
      * @param duration
      *            the value to set
      */
     public void setDuration(int duration) {
-        setDuration((double) duration);
+        getElement().setProperty("duration", (double) duration);
     }
 
     /**
-     * <p>
-     * Description copied from corresponding location in WebComponent:
-     * </p>
      * <p>
      * The duration in milliseconds to show the notification. Set to {@code 0}
      * or a negative number to disable the notification auto-closing.
@@ -521,7 +545,7 @@ public class Notification extends GeneratedVaadinNotification<Notification>
      * @return the {@code duration} property from the webcomponent
      */
     public int getDuration() {
-        return (int) getDurationDouble();
+        return (int) getElement().getProperty("duration", 0.0);
     }
 
     /**
@@ -547,30 +571,6 @@ public class Notification extends GeneratedVaadinNotification<Notification>
     public Registration addDetachListener(
             ComponentEventListener<DetachEvent> listener) {
         return super.addDetachListener(listener);
-    }
-
-    /**
-     * Adds theme variants to the component.
-     *
-     * @param variants
-     *            theme variants to add
-     */
-    public void addThemeVariants(NotificationVariant... variants) {
-        getThemeNames().addAll(
-                Stream.of(variants).map(NotificationVariant::getVariantName)
-                        .collect(Collectors.toList()));
-    }
-
-    /**
-     * Removes theme variants from the component.
-     *
-     * @param variants
-     *            theme variants to remove
-     */
-    public void removeThemeVariants(NotificationVariant... variants) {
-        getThemeNames().removeAll(
-                Stream.of(variants).map(NotificationVariant::getVariantName)
-                        .collect(Collectors.toList()));
     }
 
     private void configureComponentRenderer() {
